+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## main / unreleased
 
 * [ENHANCEMENT] New parameter log.format allows to set logging format to logfmt (default) or json (new). #184
+* [ENHANCEMENT] Add a 5 minute timeout to requests to Pods and to the Kubernetes control plane.
 
 ## v0.21.0