+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## main / unreleased
 
+## v0.22.0
+
 * [ENHANCEMENT] New parameter log.format allows to set logging format to logfmt (default) or json (new). #184
 * [ENHANCEMENT] Add a 5 minute timeout to requests to Pods and to the Kubernetes control plane. #186